Fix undefined behavior when left shifting:
../src/udp.c:47:54: runtime error: left shift of 2209677876 by 32 places cannot be represented in type 'long long'
&& ../src/udp.c:63

+++ src/udp.c
@@ -44,7 +44,7 @@ uint64_t timestampNTP_u64(void)
 	t /= 1000000000;
 	// There is 70 years (incl. 17 leap ones) offset to the Unix Epoch.
 	// No leap seconds during that period since they were not invented yet.
-	t |= ((70LL * 365 + 17) * 24 * 60 * 60 + ts.tv_sec) << 32;
+	t |= (uint64_t)((70LL * 365 + 17) * 24 * 60 * 60 + ts.tv_sec) << 32;
 	return t; // nanoseconds (technically, 232.831 picosecond units)
 }
 
@@ -60,7 +60,7 @@ uint64_t timestampNTP_RTC_u64(void) {
 	t /= 1000000000;
 	// There is 70 years (incl. 17 leap ones) offset to the Unix Epoch.
 	// No leap seconds during that period since they were not invented yet.
-	t |= ((70LL * 365 + 17) * 24 * 60 * 60 + ts.tv_sec) << 32;
+	t |= (uint64_t)((70LL * 365 + 17) * 24 * 60 * 60 + ts.tv_sec) << 32;
 	return t;
 }
